Inline compat controller layer
